[FIX] mail_activity: don't drop value for user if no (default) team is provided
mail.activity's `create` checks if a value for `team_id` is being passed, and
in that case drops values for `user_id`. However, a value for `team_id` will
always be present in the create values because there is a default method for
the field.

In case all teams have been deleted, or if team selection is restricted as per
customization, the value for `team_id` will still be present, but empty. In
that case, we don't want the value for `user_id` to be dropped silently from
the values.

Co-authored-by: CRogos <1799080+CRogos@users.noreply.github.com>